// deploy/restpe-contabilidad-front-end/src/app/modules/rrhh/domain/models/generacion-archivos.entity.ts
export interface GeneracionArchivosEntity {
  archivo_reg_codigo: string;
  archivo_reg_periodo: string;
  archivo_reg_sucursal: string;
  archivo_reg_nombre: string;
  archivo_reg_tipo_formato: string;
  archivo_reg_fecha_emision: string;
  archivo_reg_estado: string;
  archivo_reg_detalle_error?: string;
}